programmers java 이모티콘 할인행사 (성공 - int casting을 제거하여 해결)
문제에 부동소수점 관련 이슈가 존재하는 듯

+++ java/programmers/150368/Solution.java
@@ -7,7 +7,7 @@ class Solution {
 
         Case(int discountRate, int price) {
             this.discountRate = discountRate;
-            this.price = (int) (price * ((100 - discountRate) / 100.0));
+            this.price = price * (100 - discountRate) / 100;
         }
     }
